XBased Converter
This source code is free to copy and use.It was  developed by c# . It contains  string to decimal and decimal to string converter functions on XBased. Also if you need to use custom character set you may use and create your custom set.(a-Z,0-9 etc.)
Length was defined to specify length of digits.
I paid attention to the algorithm's quick operation.
There is some improvements.